requirements.txt
# Add your Python dependencies here
mcp
python-dotenv
mcp[cli]
httpx
openai